Binding constraint RHS scenarization (#1219)
* Extract methode

* Remove useless variable

* Break some cyclical include dependencies

* Add new unit test for binding constraints

* BindingConstraint group property: RED

* BindingConstraint group property: GREEN

* Refactor test: generate file in TEMP instead of datafiles

* BoundingConstraints load group: RED

* BoundingConstraints load group: GREEN

* BoundingConstraints load RHS properly

* Store TS Numbers

* Save scenario builder for BC

* Save scenario builder for BC with all rules

* Use matrix type to save time series numbers

* Properly load time series numbers

* Fix store time series numbers

* Add test store timeseries number to unit tests

* Renaming constraint -> bindingConstraint

* Update for BCTimeSeries [FAIL]

* Load BC TimeSeries [PASS]

* Test that reading different TS size fail [PASS]

* Test that all BC have the same type [PASS]

* Rename "series" member of DataSeries types as "time_series"

* Fix compilation for all plateformes

* Simple test binding constraint use time series [FAIL]

* Use vector instead of c-style array for binding constraints RTI

* Fix resizeAllTimeseriesNumbers

* Simple test binding constraint use time series [PASS]

* Rename upper_bound/lower_bound time series as greater_than/lesser_than to avoid confusion

* Simple test binding constraint use time series [PASS]

* Time series are scenarised by BindingConstraint not groups

* WIP Time series in BC

* Only load timeseries for the bionding constraint operator

* Use time_series and ts_numbers for simulation

* Handle Daily BC

* Binding constraints weekly

* Refactor test

* Lesser and Greater inequality binding constraints

* Refactor: squash a switch case split in two

* Rename time_series -> time_series_numbers for structure in BindingConstraintsList

* Binding Constraints works with multiple time series

* Binding Constraints works with several years only one TS

* Binding Constraints works with several years only one TS

* Load time_series empty files as a column of 0s

* [DEV] Add legacy loading function for time series

* Fix for 1 TS several years

* Empty TS file is an error (BC not loaded)

* [GIT] Modify githook regex to allow capitals

* [FIX] Add intermediate matrix

* RNG ts numbers for BC

* Handle range type

* Handle range type

* Use shared_ptr instead of raw pointers for binding constraints

* f

* Code clean up

* Extract methode "SeparateValue"

Remove duplication to read Link or cluster values while parsing binding constraints

* Fix ui compilation

* Rename to StoreTimeSeriesNumbersIntoOuput to be accurate

* Remove unimplemented function to store BC TSNumbers

* Initialize TSNumbers so that they exist for sizing and filling

* Properly set TSNumbers from RNG.

* Assert when trying to load wrong operator BC for study pre-860

* Remove dead code

* Extract code to save bBindingCosntraint in dedicated class

* Constness and remove "not" for "!"

* Refactoring: minor sonar issues

* Move time series loading into BindingConstraintLoader

* Refactor building file_name to save/load time series

* Code cleanup: DeadCode, Sonar issues, etc.

* Update memoryUsage

* Remove dead code BindingConstraintTimeSeries

* Minor fix

* Remove nbTimeSeriesBindingConstraints in parameters

* Fix post rebase

* Remove TODO

* Binding constraints are never used with TS-generator

* time_series -> TimeSeries

* Fix warnings

* Remove return value of "ForceReload"

This value is never used

* Update constraintBuilder

* Update ensureDataAreLoadedForAllBindingConstraints

* Check all time series in a group have the same width

* Simplification


* Simplify InitializeTSNumbers

* f time series count

* f time series count

* Remove dead code

* Reintroduce code wrongly deleted in constraintBuilder

* Minor fix


* Fix virtual issues in TimeSeriesNumbersWriter

* Add log in case loadTimeSeries fail

* Fix compilation error

* Remove bindingConstraint value from timeSeries enum

Only used for ts generator and intramodal.

* Don't clear Study::bindingConstraints in StudyRuntimeInfos::initializeBindingConstraints

* Class Renaming TimeSeriesNumbersWriter -> BindingConstraintsTimeSeriesNumbersWriter

* Fix typo

* Pluralize vector variable name

* Separate members and functions

* Rename variables for clarity

* Remove duplication

* Remove commented code

* Rename section

* Fix windows compilation issue

* Fix warning message on conversions issue uint/size_t/singed int

* Reorganize end to end tests

* Fix windows compilation

* Use macro to import win32sock transparently

* Legacy binding constraints end at 8.7

* Fix not passing parameter properly for legacy studies

* Remove left over code

* Fix opLess case when loading legacy BCs

* Resize timeseries when loading legacy BCs

* Use resize instead of reserve, fix size

* Test for legacy studies

* Handle legacy studies

Use "id" instead of name for time_series files to match legacy behaviour

* Code formatting

* Rename timeSeries -> RHSTimeSeries

* Disable kirchhof tests for now

* [sc-builder] Don't draw BC timeseries number if only one TS (#1336)

* Study on stack generate overflow, declare on head instead

* Disable tests

* Update version 8.7.0-RC1

* Set Beta1

* Invalidate cache by changing key, BETA++

* Fix version parser

* BETA++

* Revert workflows

* Give same name and id to both constraints to ensure they are processed in the same order as before

* Fix unit test, less brittle to binding constraint list order

* Fix typo

* Remove deadcode

* Fix indent

* Rename map TimeSeriesNumbers -> GroupToTimeSeriesNumbers

* When loading legacy study, place them in the same group

---------

+++ src/solver/variable/economy/profitByPlant.h
@@ -213,7 +213,7 @@ public:
         NextType::initializeFromArea(study, area);
     }
 
-    uint getMaxNumberColumns() const
+    size_t getMaxNumberColumns() const
     {
         return pNbClustersOfArea * ResultsType::count;
     }
@@ -350,7 +350,7 @@ public:
 private:
     //! Intermediate values for each year
     typename VCardType::IntermediateValuesType pValuesForTheCurrentYear;
-    unsigned int pNbClustersOfArea;
+    size_t pNbClustersOfArea;
     unsigned int pNbYearsParallel;
 
 }; // class
